Update the naming scheme for 2023
The current naming scheme isn't ideal as it lacks the authors name but
we'll use the same scheme as on the website for now (the IDs are
generated automatically based on the Pretalx data and our Python
scripts). Maybe we can improve it a bit later but it'll probably be too
difficult to keep it backwards compatible.